chore: move server functions to db/utils
